+++ CHANGES.rst
@@ -17,6 +17,11 @@ Changelog
 - Do not fail in `vocabularies.PMUsers` when `user_id` contains special chars,
   it may be the case when using `LDAP`, ignore these values.
   [gbastien]
+- Optimized `utils.sendMailIfRelevant` to not send an email several times to
+  same address.  It was only done in `MeetingItem._sendMailToGroupMembers`.
+  Removed `MeetingItem._sendMailToGroupMembers` and manage it using new
+  parameter `isGroupIds=True` in `utils.sendMailIfRelevant`.
+  [gbastien]
 - Make the `quick edit RichText field` work for DX content types :
   - added `PMRichTextWidget` useable in DX schema;
   - renamed `utils.setFieldFromAjax` to `utils.set_field_from_ajax`;